\name{ycinterextra-package}
\alias{ycinterextra-package}
\alias{ycinterextra}
\docType{package}
\title{Yield curve or zero-coupon prices interpolation and extrapolation}
\description{
Yield curve or zero-coupon prices interpolation and extrapolation using the Nelson-Siegel, Svensson, Smith-Wilson models, and Hermite cubic splines.}
\details{
\tabular{ll}{
Package: \tab ycinterextra\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-12-18\cr
License: \tab GPL-2 | GPL-3\cr
}

}
\author{
Thierry Moudiki
Maintainer: <thierry.moudiki@gmail.com>
}
\references{
Bolder, D. and Streliski, D. (1999). Yield curve modelling at the bank of canada.
Available at SSRN 1082845. \cr

CEIOPS (2010). Qis 5 risk-free interest rates extrapolation method. Technical
report, CEIOPS. \cr

FINANSTILSYNET (2010). A technical note on the Smith-Wilson method. \cr

Gilli, M., Grosse, S., and Schumann, E. (2010). Calibrating the Nelson-Siegel Svensson model. Available at SSRN 1676747. \cr

Moudiki, T. (2013). mcGlobaloptim : Global optimization using Monte Carlo and
Quasi Monte Carlo simulation. R package version 0.1. Available on CRAN. \cr

Nelson, C. R. and Siegel, A. F. (1987). Parsimonious modeling of yield curves.
Journal of Business, pages 473-489. \cr

Smith, A. and Wilson, T. (2001). Fitting yield curves with long term
constraints. Technical report, Bacon & Woodrow. Research Notes, Bacon
and Woodrow. \cr

Svensson, L. E. (1995). Estimating forward interest rates with the extended
Nelson & Siegel method. Sveriges Riksbank Quarterly Review, 3(1) :13-26. \cr
}

